+++ history.md
@@ -1,5 +1,5 @@
 ## 1.15.1 - TBD
-* Added response interceptor
+* Added ResponseInterceptor & ResponseInterceptors which can intercept responses prior to them being processed.  However they can not mutate the response at this time.  V2 will allow for full mutation of request and response in a more elegant way.
 
 ## 1.14.1 - 2017-06-30
 * The Messages class has been removed in favour of inline exception messages.  This is a backwards breaking change...but we're still pretty fluid on those things for now.  If you were importing our messages...we'd like ot know why!